Eliminate the remaining code warnings in the touched Silabs path
The dual-SDK work now passes remote build gates, so this pass tightens the changed code by removing genuine local warnings instead of leaving them behind. It fixes incorrect printf specifiers for hal_nvm_status_t, makes the reset function honor its noreturn contract, and removes dead Series-1 GPIO/poll-control leftovers that the compiler was flagging repeatedly in both Gecko SDK and Simplicity SDK builds.

This intentionally does not chase SDK-generated PTI/ZAP warnings that originate outside the maintained source tree.

Constraint: User requested minimal modifications but also no known bugs left in the MG13 support path
Rejected: Leave compiler warnings untouched because the build already passes | avoidable warnings in maintained code are unnecessary risk
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep the maintained Silabs path warning-clean even when SDK-generated code still emits unrelated warnings
Tested: git diff --check
Tested: prior GitHub Actions run 23995267520 proved the surrounding dual-SDK build path is green before this warning cleanup
Not-tested: Follow-up GitHub Actions rerun after push

+++ src/device_config/config_nv.c
@@ -31,8 +31,8 @@ void device_config_write_to_nv() {
 
     if (st != HAL_NVM_SUCCESS) {
         printf(
-            "Failed to write DEVICE_CONFIG_DATA to NV, status: %d. (bytes: %d)\r\n",
-            st, device_config_str.size);
+            "Failed to write DEVICE_CONFIG_DATA to NV, status: %lu. (bytes: %d)\r\n",
+            (unsigned long)st, device_config_str.size);
     } else {
         printf("success!\r\n");
     }
@@ -46,8 +46,8 @@ void device_config_read_from_nv() {
 
     if (st != HAL_NVM_SUCCESS) {
         printf("Failed to read NV_ITEM_DEVICE_CONFIG, using default config "
-               "instead, status: %d. (bytes: %d)\r\n",
-               st, device_config_str.size);
+               "instead, status: %lu. (bytes: %d)\r\n",
+               (unsigned long)st, device_config_str.size);
         memcpy(device_config_str.data, default_config_data,
                sizeof(default_config_data));
         device_config_str.size = strlen((const char *)default_config_data);
